Set problems

+++ Plotter/Fonts/FontChar.cs
@@ -5,13 +5,13 @@
     /// </summary>
     public struct FontChar
     {
-        public int ID         { get; set; }
-        public int X          { get; set; }
-        public int Y          { get; set; }
-        public int Width      { get; set; }
-        public int Height     { get; set; }
-        public int XOffset    { get; set; }
-        public int YOffset    { get; set; }
-        public int XAdvance   { get; set; }
+        public int   ID         { get; set; }
+        public int   X          { get; set; }
+        public int   Y          { get; set; }
+        public int   Width      { get; set; }
+        public float Height     { get; set; }
+        public float XOffset    { get; set; }
+        public float YOffset    { get; set; }
+        public float XAdvance   { get; set; }
     }
 }
